Refactor substation diagram viewer
